+++ lib/main.dart
@@ -1,14 +1,16 @@
+import 'dart:io';
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:lux/app.dart';
+import 'package:lux/const/const.dart';
 import 'package:lux/core_config.dart';
 import 'package:lux/notifier.dart';
 import 'package:lux/tr.dart';
 import 'package:lux/utils.dart';
 import 'package:window_manager/window_manager.dart';
 
-void main(args) async {
+void main(List<String> args) async {
   WidgetsFlutterBinding.ensureInitialized();
   await notifier.ensureInitialized();
 
@@ -28,7 +30,13 @@ void main(args) async {
 
     windowManager.waitUntilReadyToShow(windowOptions, () async {
       windowManager.center();
-      windowManager.show();
+      var isLaunchFromStartUp =
+          Platform.isWindows && args.contains(launchFromStartupArg);
+      if (!isLaunchFromStartUp) {
+        windowManager.show();
+      } else {
+        notifier.show(tr().launchAtStartUpMessage);
+      }
     });
 
     var isDarkMode = await readTheme() == ThemeType.dark;