+++ lib/utils.dart
@@ -5,6 +5,7 @@ import 'package:dio/dio.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:intl/intl.dart';
 import 'package:launch_at_startup/launch_at_startup.dart';
+import 'package:lux/const/const.dart';
 import 'package:lux/core_config.dart';
 import 'package:lux/core_manager.dart';
 import 'package:lux/notifier.dart';
@@ -58,6 +59,7 @@ Future<void> setAutoLaunch(CoreManager? coreManager) async {
     launchAtStartup.setup(
       appName: packageInfo.appName,
       appPath: Platform.resolvedExecutable,
+      args: [launchFromStartupArg],
     );
     var isEnabled = await launchAtStartup.isEnabled();
     if (isAutoLaunch && !isEnabled) {
@@ -145,7 +147,6 @@ Future<void> checkForUpdate() async {
 }
 
 String formatBytes(int bytes) {
-  if (bytes < 1024) return '$bytes B';
   if (bytes < 1024 * 1024) return '${(bytes / 1024).toStringAsFixed(1)} KB';
   if (bytes < 1024 * 1024 * 1024) {
     return '${(bytes / (1024 * 1024)).toStringAsFixed(1)} MB';